requirements.txt
# Core Dependencies (Latest Stable)
python-dotenv==1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0

# LLM and Agent Framework (v1.x LTS Versions)
langchain>=1.2.0
langchain-core>=0.3.81
langchain-openai>=1.1.6
langchain-anthropic>=1.3.0
langgraph>=1.0.1

# Database
psycopg2-binary==2.9.10
sqlalchemy>=2.0.36

# Vector Database
chromadb>=0.5.23
# Use dedicated integration package for better compatibility
langchain-chroma>=1.1.0

# Web Framework
streamlit>=1.52.2
streamlit-chat==0.1.1

# Data Processing
pandas>=2.2.3
numpy<2.0.0  # Staying on 1.x as requested for compatibility

# Utilities
python-dateutil==2.9.0.post0
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.1

# Scheduling & Email (Workflow A - Autonomous Monitoring)
apscheduler>=3.10.4
resend>=2.0.0

# Testing
pytest>=8.3.4
pytest-asyncio>=0.24.0
